# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login/start": {
      "post": {
        "operationId": "AuthController_startLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginStartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login next step resolved",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Request successful",
                    "data": {
                      "nextStep": "OTP_REQUIRED",
                      "challengeId": "d0f5ec4c-7d04-4bb0-b481-efc2cd981a63",
                      "otpLength": 4,
                      "phone": "9876543210",
                      "message": "OTP sent successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "statusCode": 400,
                    "message": "Error message",
                    "errorCode": "INVALID_CREDENTIALS",
                    "requestId": "6f1c2c8e-5d1b-4a8e-9f0e-0f2a1b3c4d5e"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Account is inactive or suspended (USER_INACTIVE)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "statusCode": 400,
                    "message": "Error message",
                    "errorCode": "INVALID_CREDENTIALS",
                    "requestId": "6f1c2c8e-5d1b-4a8e-9f0e-0f2a1b3c4d5e"
                  }
                }
              }
            }
          }
        },
        "summary": "Start phone-first login and determine next step",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login/password": {
      "post": {
        "operationId": "AuthController_verifyPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password accepted and OTP challenge created",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Request successful",
                    "data": {
                      "nextStep": "OTP_REQUIRED",
                      "challengeId": "d0f5ec4c-7d04-4bb0-b481-efc2cd981a63",
                      "otpLength": 4,
                      "phone": "9876543210",
                      "message": "OTP sent successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "statusCode": 400,
                    "message": "Error message",
                    "errorCode": "INVALID_CREDENTIALS",
                    "requestId": "6f1c2c8e-5d1b-4a8e-9f0e-0f2a1b3c4d5e"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Member account (PASSWORD_LOGIN_NOT_ALLOWED_FOR_MEMBER), or correct password for an inactive account (USER_INACTIVE)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "statusCode": 400,
                    "message": "Error message",
                    "errorCode": "INVALID_CREDENTIALS",
                    "requestId": "6f1c2c8e-5d1b-4a8e-9f0e-0f2a1b3c4d5e"
                  }
                }
              }
            }
          }
        },
        "summary": "Verify password for privileged phone login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified and tokens issued",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AuthResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid OTP or challenge",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "statusCode": 400,
                    "message": "Error message",
                    "errorCode": "INVALID_CREDENTIALS",
                    "requestId": "6f1c2c8e-5d1b-4a8e-9f0e-0f2a1b3c4d5e"
                  }
                }
              }
            }
          }
        },
        "summary": "Verify OTP challenge and issue tokens",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Each refresh token works once. Reusing an old one signs the session out (errorCode SESSION_REVOKED). A deactivated or suspended account gets 403 USER_INACTIVE and its session is ended.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refresh successful",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AuthResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Rotate the refresh token and issue a new access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Always succeeds with data null, also for an invalid or already ended refresh token.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Logged out successfully",
                    "data": null
                  }
                }
              }
            }
          }
        },
        "summary": "Invalidate a refresh token session",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All sessions revoked",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Request successful",
                    "data": {
                      "revokedSessions": 3
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sign out every session of the current user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password/change": {
      "post": {
        "description": "Requires the current password. Other sessions are signed out; the current one stays valid. Members have no password (403 PASSWORD_CHANGE_NOT_ALLOWED).",
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Request successful",
                    "data": {
                      "passwordChanged": true,
                      "otherSessionsSignedOut": 1
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Current password is incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "statusCode": 400,
                    "message": "Error message",
                    "errorCode": "INVALID_CREDENTIALS",
                    "requestId": "6f1c2c8e-5d1b-4a8e-9f0e-0f2a1b3c4d5e"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change the password of an imam, committee member or admin",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AuthUserDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Access token is missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "statusCode": 400,
                    "message": "Error message",
                    "errorCode": "INVALID_CREDENTIALS",
                    "requestId": "6f1c2c8e-5d1b-4a8e-9f0e-0f2a1b3c4d5e"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get currently authenticated user profile",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Masjid Core API",
    "description": "Backend for the masjid management system. Every masjid route needs a permission; see src/access/permissions.ts.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token from /auth/login/verify-otp"
      }
    },
    "schemas": {
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Request successful"
          },
          "data": {
            "type": "object",
            "description": "Payload returned by the endpoint",
            "example": {
              "id": "usr_123",
              "fullName": "Alex Johnson"
            }
          }
        },
        "required": [
          "success",
          "message",
          "data"
        ]
      },
...
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "usr_01HXYZ123"
          },
          "fullName": {
            "type": "string",
            "example": "Alex Johnson"
          },
          "email": {
            "type": "object",
            "example": "alex@platformcore.dev",
            "nullable": true
          },
          "phone": {
            "type": "object",
            "example": "+14155550123",
            "nullable": true
          },
          "status": {
            "type": "string",
            "example": "ACTIVE"
          },
          "masjidId": {
            "type": "object",
            "example": "7f2f58ad-02a7-43fb-9d7d-384047d41f25",
            "nullable": true
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": true
          },
          "isPhoneVerified": {
            "type": "boolean",
            "example": true
          },
          "isFamilyHead": {
            "type": "boolean",
            "example": false,
            "description": "Head of a family: owes the monthly imam salary share"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-19T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-19T00:00:00.000Z"
          },
          "roles": {
            "example": [
              "SUPER_ADMIN"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permissions": {
            "example": [
              "users.read",
              "users.update"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "fullName",
          "email",
          "phone",
          "status",
          "masjidId",
          "isEmailVerified",
          "isPhoneVerified",
          "isFamilyHead",
          "createdAt",
          "updatedAt",
          "roles",
          "permissions"
        ]
      },
      "AuthTokensDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.access.token"
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh.token"
          },
          "accessTokenExpiresIn": {
            "type": "number",
            "example": 900
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "accessTokenExpiresIn"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          },
          "tokens": {
            "$ref": "#/components/schemas/AuthTokensDto"
          }
        },
        "required": [
          "user",
          "tokens"
        ]
      },
      "LoginStartDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "Phone number used to start the login flow",
            "minLength": 6,
            "maxLength": 20
          }
        },
        "required": [
          "phone"
        ]
      },
      "LoginPasswordDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "Phone number used to continue the login flow",
            "minLength": 6,
            "maxLength": 20
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!",
            "format": "password"
          }
        },
        "required": [
          "phone",
          "password"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "Phone number associated with the OTP challenge",
            "minLength": 6,
            "maxLength": 20
          },
          "challengeId": {
            "type": "string",
            "example": "d0f5ec4c-7d04-4bb0-b481-efc2cd981a63"
          },
          "otp": {
            "type": "string",
            "example": "1111",
            "description": "Digits only. The expected length is returned as otpLength when the challenge is created.",
            "minLength": 4,
            "maxLength": 8
          }
        },
        "required": [
          "phone",
          "challengeId",
          "otp"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh.token",
            "description": "Valid refresh token from a previous authentication response"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "12345678",
            "format": "password"
          },
          "newPassword": {
            "type": "string",
            "example": "a-new-strong-password",
            "format": "password",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
    }
  }
}